Cut package scope: drop CLI and discovery, move storage to pipeline/
The installable package now contains only the library — dataset
client + scrapers. Ops code stays in the repo but out of the wheel.

- Delete ats_scrapers.cli and the console entry point. Three of four
  commands were demos or operator-only; list_ats() is now exported
  from the package root to replace `list-ats`.
- Delete ats_scrapers.discovery — a docstring-only stub referencing
  root-level modules that no longer exist — and the discovery extra.
- Move ats_scrapers.storage (R2 client + DatasetPublisher) to the
  repo-only pipeline/ directory. Its dependencies move from the
  publish extra to pipeline/requirements.txt; CI installs both. The
  publish-ats-companies workflow is untouched (its script was already
  standalone).
- The `all` extra is now [scrapers,parquet]. Tests are no longer
  shipped in the sdist since the suite covers repo-only code.
