+++ pipeline/__init__.py
@@ -1,0 +1,6 @@
+"""Ops pipeline — Cloudflare R2 client and dataset publisher (not shipped on PyPI)."""
+
+from pipeline.publisher import DatasetPublisher, PublishResult
+from pipeline.r2 import R2Client, R2Config
+
+__all__ = ["DatasetPublisher", "PublishResult", "R2Client", "R2Config"]
