+++ pipeline/requirements.txt
@@ -1,0 +1,27 @@
+# Dependencies for the repo-only ops pipeline (pipeline/ and scripts/).
+# Not part of the installable ats-scrapers package — install with:
+#   pip install -r pipeline/requirements.txt
+boto3>=1.34
+pyarrow>=15.0
+# Cross-ATS dedup operates on a 3M+-row keys frame and writes the
+# global all.parquet. Polars keeps the working set Arrow-backed so peak
+# RSS stays below 1 GB on a 7.6 GB box; pandas equivalents peaked at
+# 5-6 GB.
+polars>=0.20
+# Cross-source fuzzy-match dedup (Phase 2): catches near-dups that the
+# exact-key passes miss because of formatting variations between
+# aggregators (eures NUTS-code locations vs Bundesagentur full text,
+# trailing parenthesized occupation codes, etc.).
+rapidfuzz>=3.0
+# markdownify is the canonical HTML→markdown converter; used by
+# scripts/normalize_descriptions.py which runs post-scrape to clean the
+# raw HTML several scrapers emit.
+markdownify>=0.13
+# linkify-it-py runs as the final pass over each normalized
+# description: bare URLs and emails left in plain-text postings get
+# wrapped in CommonMark autolinks so the viewer renders them clickable.
+linkify-it-py>=2.0
+# zstandard backs the optional compress=True path on the persistent
+# DescriptionCache (scripts/run_pipeline.py). Workday's cache uses it
+# to keep ~700k entries under 3 GB on disk.
+zstandard>=0.22
