--- src/ats_scrapers/discovery/__init__.py
@@ -1,7 +1,0 @@
-"""Discovery — find new companies on each ATS.
-
-Backed by SearXNG (self-hosted, unlimited), SerpAPI, or Firecrawl. The legacy
-implementations live at the repo root in `searxng_discovery.py`,
-`serpapi_discovery.py`, and `firecrawl_discovery.py`. This module will wrap
-them progressively as scrapers are ported.
-"""
