--- src/ats_scrapers/storage/__init__.py
@@ -1,6 +1,0 @@
-"""Storage layer — Cloudflare R2 client and dataset publisher."""
-
-from ats_scrapers.storage.publisher import DatasetPublisher, PublishResult
-from ats_scrapers.storage.r2 import R2Client, R2Config
-
-__all__ = ["DatasetPublisher", "PublishResult", "R2Client", "R2Config"]
